Move combine helpers to new file
