+++ EssentialApp/EssentialApp.xcodeproj/xcuserdata/prabhattiwari.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -3,7 +3,13 @@
 <plist version="1.0">
 <dict>
 	<key>SchemeUserState</key>
-	<dict/>
+	<dict>
+		<key>EssentialApp.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>5</integer>
+		</dict>
+	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
 		<key>4306AF6E2826055900985AF5</key>
